Split tee profile worker concurrency tests (#119)
Move concurrency scaffolding into one private support module and keep the
reentrancy and concurrent-worker suites focused on their separate concerns.
Update the test-helper lint classification and wheel-layout snapshot for the
renamed modules.
